Code cleanup and provide a provision for labeling unlabeled data with predicitons which can then be vetted by a user, and added to the training data.

+++ ring_camera_convnet.py
@@ -1,16 +1,14 @@
 import tensorflow as tf
 import numpy as np
-import random
+import math, random
+import os, shutil, glob
 import csv
-import os
-import shutil
 import cv2
-import keras
-import glob
+
 from PIL import Image
 from pathlib import Path
 
-weights_filename = "ring_convnet_weights/checkpt.ckpt"
+model_filename = "ring_convnet_weights/checkpt.ckpt"
 np_train_images_file = 'train_images.npy'
 np_train_labels_file = 'train_labels.npy'
 np_val_images_file = 'val_images.npy'
@@ -27,7 +25,7 @@ all_numpy_files = [
 ]
 
 use_cpu = False
-max_images = 1000
+max_images = 15000
 
 labels_to_consider = ['none', 'car', 'dog', 'turkey', 'deer']
 
@@ -57,7 +55,8 @@ def get_img_dict_from_csv(csv_path):
 
 def save_img_dict_to_csv(img_dict, csv_path):
     # copy just in case we've messed something up...
-    shutil.copy(csv_path, str(csv_path) + '.bak')
+    if os.path.exists(csv_path):
+        shutil.copy(csv_path, str(csv_path) + '.bak')
 
     # overwrites any pre-existing file
     with open(os.path.abspath(csv_path), 'w', newline='') as csvfile:
@@ -201,21 +200,21 @@ def split_data_into_groups_bucketize(img_dict, max_images=None):
     return train_imgs, train_labels, val_imgs, val_labels, test_imgs, test_labels
 
 def get_compiled_model(img_shape, num_outputs):
-    inputs = keras.Input(shape=img_shape)
+    inputs = tf.keras.Input(shape=img_shape)
 
-    x = keras.layers.Conv2D(filters=32, kernel_size=3, activation='relu')(inputs)
-    x = keras.layers.MaxPool2D(pool_size=2)(x)
-    x = keras.layers.Conv2D(filters=64, kernel_size=3, activation='relu')(x)
-    x = keras.layers.MaxPool2D(pool_size=2)(x)
-    x = keras.layers.Conv2D(filters=128, kernel_size=3, activation='relu')(x)
-    x = keras.layers.Dropout(0.5)(x)
-    x = keras.layers.Flatten()(x)
-    outputs = keras.layers.Dense(num_outputs, activation='softmax')(x)
+    x = tf.keras.layers.Conv2D(filters=32, kernel_size=3, activation='relu')(inputs)
+    x = tf.keras.layers.MaxPool2D(pool_size=2)(x)
+    x = tf.keras.layers.Conv2D(filters=64, kernel_size=3, activation='relu')(x)
+    x = tf.keras.layers.MaxPool2D(pool_size=2)(x)
+    x = tf.keras.layers.Conv2D(filters=128, kernel_size=3, activation='relu')(x)
+    x = tf.keras.layers.Dropout(0.5)(x)
+    x = tf.keras.layers.Flatten()(x)
+    outputs = tf.keras.layers.Dense(num_outputs, activation='softmax')(x)
 
-    model = keras.Model(inputs=inputs, outputs=outputs)
+    model = tf.keras.Model(inputs=inputs, outputs=outputs)
 
     model.compile(
-        optimizer=keras.optimizers.RMSprop(learning_rate=0.001),
+        optimizer=tf.keras.optimizers.RMSprop(learning_rate=0.001),
         loss='sparse_categorical_crossentropy',
         metrics=['accuracy']
     )
@@ -301,21 +300,21 @@ def train_and_evaluate(force_reload_images):
             validation_data=(val_imgs, val_labels),
             epochs=100,
             callbacks=[
-                keras.callbacks.EarlyStopping(patience=15),
-                keras.callbacks.ModelCheckpoint(
-                    filepath=weights_filename,
+                tf.keras.callbacks.EarlyStopping(patience=15),
+                tf.keras.callbacks.ModelCheckpoint(
+                    filepath=model_filename,
                     save_weights_only=False,
                     save_best_only=True,
                     monitor='val_accuracy',
                     mode='max'
                 ),
-                keras.callbacks.ReduceLROnPlateau(
+                tf.keras.callbacks.ReduceLROnPlateau(
                     monitor='val_accuracy',
                     factor=0.5,
                     patience=6,
                     min_lr=0.00005
                 ),
-                keras.callbacks.TensorBoard(
+                tf.keras.callbacks.TensorBoard(
                     log_dir=os.path.abspath("./tensorboard_logs")
                 )
             ],
@@ -327,7 +326,7 @@ def train_and_evaluate(force_reload_images):
 
     # load in the weights from the epoch with the maximum accuracy
     # and evaluate the model on the test data
-    model.load_weights(weights_filename)
+    model.load_weights(model_filename)
     loss, accuracy = model.evaluate(
         test_imgs,
         test_labels,
@@ -344,11 +343,10 @@ def evaluate_only(show_predict_loop=False):
     # what's the image shape for all images in this dataset
     img_shape = train_imgs[0].shape
 
-    model = get_compiled_model(img_shape=img_shape, num_outputs=num_labels)
+    # load the model from the best epoch checkpoint
+    model = tf.keras.models.load_model(model_filename)
     print(model.summary())
 
-    # evaluate the model from the best epoch checkpoint
-    model.load_weights(weights_filename)
     model.evaluate(
         test_imgs,
         test_labels,
@@ -380,6 +378,42 @@ def evaluate_only(show_predict_loop=False):
                 break
             cv2.destroyAllWindows()
 
+def create_predictions_on_unlabeled_data(csv_path, img_folder):
+    prediction_batch_size = 1000
+
+    labeled_imgs = get_img_dict_from_csv(csv_path)
+    all_img_paths = get_all_images_in_folder(img_folder)
+    all_img_paths = list(filter(lambda x: x not in labeled_imgs, all_img_paths))
+    total_imgs_to_predict = len(all_img_paths)
+
+    # load the model from the best epoch checkpoint
+    model = tf.keras.models.load_model(model_filename)
+    print(model.summary())
+
+    output_dict = {}
+    label_dict = {k: v for k, v in enumerate(labels_to_consider)}
+
+    # make predicitons on batches so we don't run out of memory!
+    total_batches = math.ceil(total_imgs_to_predict / prediction_batch_size)
+    for batch_num in range(0, total_batches):
+        print(f'Predicting batch [{batch_num + 1} / {total_batches}]...')
+        start_idx = batch_num * prediction_batch_size
+        end_idx = (batch_num + 1) * prediction_batch_size
+        batch_img_paths = all_img_paths[start_idx:end_idx]
+        batch_imgs = list(map(convert_img_to_tensor, batch_img_paths))
+        batch_imgs = np.array(batch_imgs)
+        predictions = model.predict(batch_imgs)
+
+        # loop over each prediction made on the test data
+        for idx, img_path in enumerate(batch_img_paths):
+            prediction = predictions[idx]
+            int_label = tf.argmax(prediction).numpy()
+            label = label_dict[int_label]
+            output_dict[img_path] = label
+
+    save_img_dict_to_csv(output_dict, 'unlabeled_imgs.csv')
+
 if __name__ == '__main__':
-    train_and_evaluate(force_reload_images=True)
-    #evaluate_only(show_predict_loop=True)
+    #train_and_evaluate(force_reload_images=True)
+    evaluate_only(show_predict_loop=True)
+    #create_predictions_on_unlabeled_data(csv_path, './ring_downloader/ring_data/sept_through_nov_2023/frames/400max')
